Export BasePlugin (#408)
BasePlugin is not described in the TypeScript types published with this module (i.e. types/index.d.ts) which means that it is difficult to write custom plugins in TypeScript or use VS Code with a jsconfig.json file with "checkJS": true.

This commit exposes BasePlugin on the Cloudant factory method so that it can be imported directly from @cloudant/cloudant.

This means that custom plugin authors do not need to require @cloudant/cloudant/plugins/base.