Defer plugin callbacks to next tick
Nock 13.x has changed to better reflect the semantics of modern Node versions.
A side effect of this was the test
`CloudantClient/#db using listeners/with ComplexPlugin2` failing with the `end` event being triggered twice, causing two done callbacks.
The changes suggest that nock moved the abort event to the next tick. This ensures our plugin callbacks also move so that abort is handled correctly.